pyproject.toml
[tool.poetry]
name = "envix"
version = "0.1.0"
description = "convinient secret manager."
authors = ["yassun7010 <yassun7010@outlook.com>"]
readme = "README.md"
packages = [{ include = "envix", from = "src" }]

[tool.poetry.scripts]
envix = "envix.cli.__main__:main"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.6.4"
rich = "^13.7.1"
rich-argparse = "^1.4.0"
pyyaml = "^6.0.1"
google-cloud-secret-manager = "^2.19.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pyright = "^1.1.357"
ruff = "^0.3.5"
taskipy = "^1.12.2"

[tool.taskipy.tasks]
format = "ruff format ."
lint = "ruff check ."
typecheck = "pyright src/** tests/**"
test = "pytest"
ci = "task format && task lint && task typecheck && task test"
schema = "envix config schema > schemas/config.json"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
